+++ src/rc_input/config.yml
@@ -22,6 +22,11 @@
 # language spec available here: https://yaml.org/spec/1.2/spec.html
 
 
+# How frequently the receiver checks for new inputs
+read interval: 50 / 1000  # seconds
+
+
+# The pins connected to channels from the physical receiver
 pins:
   RUDDER:
     pin_name: "P0_0"
