+++ src/rc_input/rc_input_thread.py
@@ -1,11 +1,9 @@
 from threading import Thread
 from time import sleep
 
-from src.rc_input.pin_config_reader import read_pin_config
+from src.rc_input.config_reader import read_pin_config, read_interval
 from src.rc_input.rc_receiver import RCReceiver
 from src.rc_input.rc_broadcaster import make_broadcaster, RCInputBroadcasterType
-
-RC_READ_INTERVAL = 50 / 1000  # 50 milliseconds
 
 
 class RCInputThread(Thread):
@@ -13,7 +11,7 @@ class RCInputThread(Thread):
 
     Should accept multiple boat configurations, and should be general enough to allow for easy extension.
     """
-    def __init__(self, config=None):
+    def __init__(self):
         """Builds a new RC input thread."""
         super().__init__()
 
@@ -21,9 +19,10 @@ class RCInputThread(Thread):
         self.receiver = RCReceiver(broadcaster=self.broadcaster, pins=read_pin_config())
 
         self.keep_reading = True
+        self.read_interval = read_interval()
 
     def run(self):
         """Starts a regular input read interval."""
         while True:
             self.receiver.send_inputs()
-            sleep(RC_READ_INTERVAL)
+            sleep(self.read_interval)
